QPID/Avro changes for prior commit. part 3 of N.

+++ trunk/sources/modules/RoboKindAvroQPIDModule/qpid-avro-cs/avro-src-1.7.0/lang/csharp/src/apache/main/IO/BinaryDecoder.cs
@@ -135,7 +135,8 @@ namespace Avro.IO
 
         public string ReadString()
         {
-            int length = ReadInt();
+            //int length = ReadInt();
+            long length = ReadLong();
             byte[] buffer = new byte[length];
             //TODO: Fix this because it's lame;
             ReadFixed(buffer);
